[ci:force] Extract StringUtils as well
