+++ lib/testing/src/lib/protractor/core/models/tenant.ts
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-import { StringUtil } from '../utils/string.util';
+import { StringUtil } from '../../../shared/utils/string.util';
 
 /**
  * Create tenant JSON Object
