+++ lib/testing/src/lib/shared/utils/string.util.ts
